spring mvc fix

+++ module-09-spring-core/ecinema/src/main/java/com/epam/example/ecinema/web/ClientController.java
@@ -3,6 +3,7 @@ package com.epam.example.ecinema.web;
 import java.util.List;
 
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -15,6 +16,7 @@ import org.springframework.web.bind.annotation.ResponseBody;
 
 import com.epam.example.ecinema.domain.Client;
 import com.epam.example.ecinema.service.ClientService;
+import com.epam.example.ecinema.web.exception.ForbiddenException;
 
 @Controller
 @RequestMapping("/clients")
@@ -38,14 +40,20 @@ public class ClientController {
 	
 	@RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
 	@ResponseBody
-	public String deleteClient(@PathVariable("id") Long id) {
+	public String deleteClient(@PathVariable("id") Long id, HttpSession session) {
+		if (session.getAttribute("loginSession") == null) {
+			throw new ForbiddenException();
+		}
 		service.removeClientById(id);
 		return "";
 	}
 	
 	@RequestMapping(value = {"", "/"}, method = RequestMethod.POST)
 	@ResponseBody
-	public Long createClient(@RequestBody MultiValueMap<String, String> body, HttpServletResponse resp) {
+	public Long createClient(@RequestBody MultiValueMap<String, String> body, HttpServletResponse resp, HttpSession session) {
+		if (session.getAttribute("loginSession") == null) {
+			throw new ForbiddenException();
+		}
 		resp.setStatus(201);
 		Client client = new Client();
 		client.setId(0L);
